处理一下文本强度参数的类型

+++ taichubackend-application/src/main/java/com/taichu/application/executor/StoryboardTextAndImgTaskExecutor.java
@@ -81,8 +81,8 @@ public class StoryboardTextAndImgTaskExecutor extends AbstractTaskExecutor {
             return Map.of();
         }
         GenerateStoryboardImgRequest storyboardImgRequest = (GenerateStoryboardImgRequest) request;
-        Double styleScale = storyboardImgRequest.getStyleScale();
-        Double scale = storyboardImgRequest.getScale();
+        Long styleScale = storyboardImgRequest.getStyleScale();
+        Long scale = storyboardImgRequest.getScale();
         String imageStyle = storyboardImgRequest.getImageStyle();
 
         Map<String, String> params = new HashMap<>();
